@@ -1,22 +1,13 @@
-import React, { useContext } from "react";
-import { Route, Redirect } from "react-router-dom";
-import AuthContext from "../../context/auth/authContext";
+import React, { useContext } from 'react';
+import { Navigate } from 'react-router-dom';
+import AuthContext from '../../context/auth/authContext';
 
-const PrivateRoute = ({ component: Component, ...rest }) => {
+const PrivateRoute = ({ component: Component }) => {
   const authContext = useContext(AuthContext);
   const { isAuthenticated, loading } = authContext;
-  return (
-    <Route
-      {...rest}
-      render={props =>
-        !isAuthenticated && !loading ? (
-          <Redirect to="login" />
-        ) : (
-          <Component {...props} />
-        )
-      }
-    />
-  );
+
+  if (isAuthenticated && !loading) return <Component />;
+  return <Navigate to='login' />;
 };
 
 export default PrivateRoute;
