@@ -1,8 +1,9 @@
-import React, { useState, useContext, useEffect } from "react";
-import AuthContext from "../../context/auth/authContext";
-import AlertContext from "../../context/alert/alertContext";
+import React, { useState, useContext, useEffect } from 'react';
+import { Navigate } from 'react-router-dom';
+import AuthContext from '../../context/auth/authContext';
+import AlertContext from '../../context/alert/alertContext';
 
-const Login = props => {
+const Login = () => {
   const alertContext = useContext(AlertContext);
   const authContext = useContext(AuthContext);
 
@@ -11,12 +12,8 @@ const Login = props => {
   const { login, error, clearErrors, isAuthenticated } = authContext;
 
   useEffect(() => {
-    if (isAuthenticated) {
-      props.history.push("/");
-    }
-
-    if (error === "User already exists") {
-      setAlert(error, "danger");
+    if (error === 'User already exists') {
+      setAlert(error, 'danger');
       clearErrors();
     }
 
@@ -24,53 +21,55 @@ const Login = props => {
   }, [error, isAuthenticated, props.history]);
 
   const [user, setUser] = useState({
-    email: "",
-    password: ""
+    email: '',
+    password: '',
   });
 
   const { email, password } = user;
 
-  const onChange = e => setUser({ ...user, [e.target.name]: e.target.value });
+  const onChange = (e) => setUser({ ...user, [e.target.name]: e.target.value });
 
-  const onSubmit = e => {
+  const onSubmit = (e) => {
     e.preventDefault();
-    if (email === "" || password === "") {
-      setAlert("Please fill in all fileds", "danger");
+    if (email === '' || password === '') {
+      setAlert('Please fill in all fileds', 'danger');
     } else {
       login({ email, password });
     }
   };
 
+  if (isAuthenticated) return <Navigate to='/' />;
+
   return (
-    <div className="form-container">
+    <div className='form-container'>
       <h1>
-        Account <span className="text-primary">Login</span>
+        Account <span className='text-primary'>Login</span>
       </h1>
       <form onSubmit={onSubmit}>
-        <div className="form-group">
-          <label htmlFor="email">Email Address</label>
+        <div className='form-group'>
+          <label htmlFor='email'>Email Address</label>
           <input
-            type="email"
-            name="email"
+            type='email'
+            name='email'
             value={email}
             onChange={onChange}
             required
           />
         </div>
-        <div className="form-group">
-          <label htmlFor="password">Password</label>
+        <div className='form-group'>
+          <label htmlFor='password'>Password</label>
           <input
-            type="password"
-            name="password"
+            type='password'
+            name='password'
             value={password}
             onChange={onChange}
             required
           />
         </div>
         <input
-          type="submit"
-          value="Login"
-          className="btn btn-primary btn-block"
+          type='submit'
+          value='Login'
+          className='btn btn-primary btn-block'
         />
       </form>
     </div>
