@@ -1,8 +1,9 @@
-import React, { useState, useContext, useEffect } from "react";
-import AlertContext from "../../context/alert/alertContext";
-import AuthContext from "../../context/auth/authContext";
+import React, { useState, useContext, useEffect } from 'react';
+import { Navigate } from 'react-router-dom';
+import AlertContext from '../../context/alert/alertContext';
+import AuthContext from '../../context/auth/authContext';
 
-const Register = props => {
+const Register = (props) => {
   const alertContext = useContext(AlertContext);
   const authContext = useContext(AuthContext);
 
@@ -11,12 +12,8 @@ const Register = props => {
   const { register, error, clearErrors, isAuthenticated } = authContext;
 
   useEffect(() => {
-    if (isAuthenticated) {
-      props.history.push("/");
-    }
-
-    if (error === "User already exists") {
-      setAlert(error, "danger");
+    if (error === 'User already exists') {
+      setAlert(error, 'danger');
       clearErrors();
     }
 
@@ -24,83 +21,85 @@ const Register = props => {
   }, [error, isAuthenticated, props.history]);
 
   const [user, setUser] = useState({
-    name: "",
-    email: "",
-    password: "",
-    password2: ""
+    name: '',
+    email: '',
+    password: '',
+    password2: '',
   });
 
   const { name, email, password, password2 } = user;
 
-  const onChange = e => setUser({ ...user, [e.target.name]: e.target.value });
+  const onChange = (e) => setUser({ ...user, [e.target.name]: e.target.value });
 
-  const onSubmit = e => {
+  const onSubmit = (e) => {
     e.preventDefault();
-    if (name === "" || email === "" || password === "") {
-      setAlert("Please enter all fields", "danger");
+    if (name === '' || email === '' || password === '') {
+      setAlert('Please enter all fields', 'danger');
     } else if (password !== password2) {
-      setAlert("Passwords do not match", "danger");
+      setAlert('Passwords do not match', 'danger');
     } else {
       register({
         name,
         email,
-        password
+        password,
       });
     }
   };
 
+  if (isAuthenticated) return <Navigate to='/' />;
+
   return (
-    <div className="form-container">
+    <div className='form-container'>
       <h1>
-        Account <span className="text-primary">Register</span>
+        Account <span className='text-primary'>Register</span>
       </h1>
       <form onSubmit={onSubmit}>
-        <div className="form-group">
-          <label htmlFor="name">Name</label>
+        <div className='form-group'>
+          <label htmlFor='name'>Name</label>
           <input
-            type="text"
-            name="name"
+            type='text'
+            name='name'
             value={name}
             onChange={onChange}
             required
           />
         </div>
-        <div className="form-group">
-          <label htmlFor="email">Email Address</label>
+        <div className='form-group'>
+          <label htmlFor='email'>Email Address</label>
           <input
-            type="email"
-            name="email"
+            type='email'
+            name='email'
             value={email}
             onChange={onChange}
             required
           />
         </div>
-        <div className="form-group">
-          <label htmlFor="password">Password</label>
+        <div className='form-group'>
+          <label htmlFor='password'>Password</label>
           <input
-            type="password"
-            name="password"
+            type='password'
+            name='password'
             value={password}
             onChange={onChange}
             required
-            minLength="6"
+            minLength='6'
           />
         </div>
-        <div className="form-group">
-          <label htmlFor="password">Confirm Password</label>
+        <div className='form-group'>
+          <label htmlFor='password'>Confirm Password</label>
           <input
-            type="password"
-            name="password2"
+            type='password'
+            name='password2'
             value={password2}
             onChange={onChange}
             required
-            minLength="6"
+            minLength='6'
           />
         </div>
         <input
-          type="submit"
-          value="Register"
-          className="btn btn-primary btn-block"
+          type='submit'
+          value='Register'
+          className='btn btn-primary btn-block'
         />
       </form>
     </div>
