@@ -2,11 +2,15 @@ import React, { Fragment } from 'react';
 import spinner from './spinner.gif';
 
 function Spinner() {
-    return (
-        <Fragment>
-            <img src={spinner} alt="Loading..." style={{width: '200px', margin: 'auto', display: 'block' }}/>
-        </Fragment>
-    )
+  return (
+    <Fragment>
+      <img
+        src={spinner}
+        alt='Loading...'
+        style={{ width: '200px', margin: 'auto', display: 'block' }}
+      />
+    </Fragment>
+  );
 }
 
-export default Spinner
+export default Spinner;
