@@ -1,5 +1,5 @@
 import React, { useReducer } from "react";
-import uuid from 'uuid';
+import { v4 as uuid } from 'uuid';
 import AlertContext from "./alertContext";
 import alertReducer from "./alertReducer";
 import {
@@ -14,7 +14,7 @@ const AlertState = props => {
 
   // Set Alert
     const setAlert = (msg, type, timeout = 5000) => {
-        const id = uuid.v4();
+        const id = uuid();
         dispatch({
             type: SET_ALERT,
             payload: { msg, type, id}
